Generalize UI Controller and Restful API Controller, Created new Class to handle common methods
